chore(widgets): Encapsulate the hacky / complex setViewState()
